remove config api from node env manager - seems not use
move createFeatureEnvironmentsMapping to cli (where it's used)

move files and imports accordingly
