+++ packages/engine-cli/src/find-features/analyze-features.ts
@@ -1,11 +1,10 @@
 import type { IFileSystemSync } from '@file-services/types';
 import { concat } from '@wixc3/common';
 import { flattenTree } from '@wixc3/engine-core';
-import type { IConfigDefinition } from '@wixc3/engine-runtime-node';
 import type { SetMultiMap } from '@wixc3/patterns';
 import { childPackagesFromContext, resolveDirectoryContext, type INpmPackage } from '@wixc3/resolve-directory-context';
 import { isFeatureFile } from './build-constants.js';
-import type { IFeatureDefinition } from '../types.js';
+import type { IConfigDefinition, IFeatureDefinition } from '../types.js';
 import { loadFeaturesFromPaths } from './load-features-from-paths.js';
 import { mergeAll, mergeResults } from './merge.js';
