+++ packages/engine-cli/src/find-features/parse-env.ts
@@ -6,7 +6,7 @@ import {
     type EnvironmentMode,
     type MultiEnvironment,
 } from '@wixc3/engine-core';
-import { type IEnvironmentDescriptor } from '@wixc3/engine-runtime-node';
+import type { IEnvironmentDescriptor } from '../types.js';
 
 const convertEnvToIEnv = (env: AnyEnvironment): IEnvironmentDescriptor => {
     const { env: name, envType: type } = env;
