+++ packages/engine-cli/src/metadata-files.ts
@@ -1,17 +1,13 @@
 import fs from 'node:fs';
 import path from 'node:path';
-import type {
-    ConfigurationEnvironmentMapping,
-    FeatureEnvironmentMapping,
-    createFeatureEnvironmentsMapping,
-} from './node-env-manager.js';
+import { ConfigurationEnvironmentMapping, FeatureEnvironmentMapping } from './types.js';
 
 export function readMetadataFiles(dir: string) {
     try {
         const envMappingFilePath = path.join(dir, 'metadata', 'engine-feature-environments-mapping.json');
-        const featureEnvironmentsMapping = JSON.parse(fs.readFileSync(envMappingFilePath, 'utf8')) as ReturnType<
-            typeof createFeatureEnvironmentsMapping
-        >;
+        const featureEnvironmentsMapping = JSON.parse(
+            fs.readFileSync(envMappingFilePath, 'utf8'),
+        ) as FeatureEnvironmentMapping;
         const engineConfigMappingFilePath = path.join(dir, 'metadata', 'engine-config-mapping.json');
         const configMapping = JSON.parse(
             fs.readFileSync(engineConfigMappingFilePath, 'utf8'),
