+++ packages/runtime-node/src/index.ts
@@ -1,13 +1,8 @@
 export * from './ipc-host.js';
 export * from './parent-port-host.js';
 export * from './ws-node-host.js';
-export * from './environments.js';
-export * from './import-modules.js';
 export * from './launch-http-server.js';
-export * from './load-top-level-config.js';
 export * from './node-env-manager.js';
-export * from './types.js';
 export * from './worker-thread-initializer2.js';
 export * from './metrics-utils.js';
 export * from './micro-rpc.js';
-export * from './metadata-files.js';
